+++ lib/cdc/event.go
@@ -22,7 +22,7 @@ type Event interface {
 	DeletePayload() bool
 	GetTableName() string
 	GetData(pkMap map[string]any, config *kafkalib.TopicConfig) (map[string]any, error)
-	GetOptionalSchema() map[string]typing.KindDetails
+	GetOptionalSchema() (map[string]typing.KindDetails, error)
 	// GetColumns will inspect the envelope's payload right now and return.
 	GetColumns() (*columns.Columns, error)
 }
