+++ lib/cdc/util/optional_schema.go
@@ -1,22 +1,27 @@
 package util
 
 import (
+	"fmt"
 	"log/slog"
 
 	"github.com/artie-labs/transfer/lib/debezium"
 	"github.com/artie-labs/transfer/lib/typing"
 )
 
-func (s *SchemaEventPayload) GetOptionalSchema() map[string]typing.KindDetails {
+func (s *SchemaEventPayload) GetOptionalSchema() (map[string]typing.KindDetails, error) {
 	fieldsObject := s.Schema.GetSchemaFromLabel(debezium.After)
 	if fieldsObject == nil {
 		// AFTER schema does not exist.
-		return nil
+		return nil, nil
 	}
 
 	schema := make(map[string]typing.KindDetails)
 	for _, field := range fieldsObject.Fields {
-		kd := field.ToKindDetails()
+		kd, err := field.ToKindDetails()
+		if err != nil {
+			return nil, fmt.Errorf("failed to convert field (%v), to kind details: %w", field, err)
+		}
+
 		if kd == typing.Invalid {
 			slog.Warn("Skipping field from optional schema b/c we cannot determine the data type", slog.String("field", field.FieldName))
 			continue
@@ -25,5 +30,5 @@ func (s *SchemaEventPayload) GetOptionalSchema() map[string]typing.KindDetails {
 		schema[field.FieldName] = kd
 	}
 
-	return schema
+	return schema, nil
 }
