+++ lib/debezium/encode/decimal.go
@@ -1,4 +1,4 @@
-package debezium
+package encode
 
 import (
 	"math/big"
