+++ lib/debezium/encode/decimal_test.go
@@ -1,4 +1,4 @@
-package debezium
+package encode
 
 import (
 	"fmt"
